modify test cases
