+++ AElf.Kernel.Tests/TxPool/TxPoolServiceTest.cs
@@ -28,7 +28,7 @@ namespace AElf.Kernel.Tests.TxPool
         }
         
         [Fact]
-        public async Task AddTxTest()
+        public async Task ServiceTest()
         {
             var pool = GetPool();
             
@@ -49,9 +49,9 @@ namespace AElf.Kernel.Tests.TxPool
             var res = await poolService.AddTxAsync(tx1);
             Assert.True(res);
             
-            Assert.Equal(1, (int)poolService.GetTmpSizeAsync().Result);
-            Assert.Equal(0, (int)poolService.GetWaitingSizeAsync().Result);
-            Assert.Equal(0, (int)poolService.GetExecutableSizeAsync().Result);
+            Assert.Equal(1, (int) await poolService.GetTmpSizeAsync());
+            Assert.Equal(0, (int) await poolService.GetWaitingSizeAsync());
+            Assert.Equal(0, (int) await poolService.GetExecutableSizeAsync());
             Assert.Equal(1, (int)pool.Size);
             
             var tx2 = new Transaction
@@ -63,9 +63,9 @@ namespace AElf.Kernel.Tests.TxPool
             res = await poolService.AddTxAsync(tx2);
             
             Assert.True(res);
-            Assert.Equal(2, (int)poolService.GetTmpSizeAsync().Result);
-            Assert.Equal(0, (int)poolService.GetWaitingSizeAsync().Result);
-            Assert.Equal(0, (int)poolService.GetExecutableSizeAsync().Result);
+            Assert.Equal(2, (int) await poolService.GetTmpSizeAsync());
+            Assert.Equal(0, (int) await poolService.GetWaitingSizeAsync());
+            Assert.Equal(0, (int) await poolService.GetExecutableSizeAsync());
             Assert.Equal(2, (int)pool.Size);
             
             
@@ -78,9 +78,9 @@ namespace AElf.Kernel.Tests.TxPool
             res = await poolService.AddTxAsync(tx3);
             
             Assert.True(res);
-            Assert.Equal(3, (int)poolService.GetTmpSizeAsync().Result);
-            Assert.Equal(0, (int)poolService.GetWaitingSizeAsync().Result);
-            Assert.Equal(0, (int)poolService.GetExecutableSizeAsync().Result);
+            Assert.Equal(3, (int) await poolService.GetTmpSizeAsync());
+            Assert.Equal(0, (int) await poolService.GetWaitingSizeAsync());
+            Assert.Equal(0, (int) await poolService.GetExecutableSizeAsync());
             Assert.Equal(3, (int)pool.Size);
             
             
@@ -92,9 +92,9 @@ namespace AElf.Kernel.Tests.TxPool
             };
             res = await poolService.AddTxAsync(tx4);
             Assert.True(res);
-            Assert.Equal(4, (int)poolService.GetTmpSizeAsync().Result);
-            Assert.Equal(0, (int)poolService.GetWaitingSizeAsync().Result);
-            Assert.Equal(0, (int)poolService.GetExecutableSizeAsync().Result);
+            Assert.Equal(4, (int) await poolService.GetTmpSizeAsync());
+            Assert.Equal(0, (int) await poolService.GetWaitingSizeAsync());
+            Assert.Equal(0, (int) await poolService.GetExecutableSizeAsync());
             Assert.Equal(4, (int)pool.Size);
             
             var tx5 = new Transaction
@@ -106,10 +106,10 @@ namespace AElf.Kernel.Tests.TxPool
             res = await poolService.AddTxAsync(tx5);
             Assert.True(res);
             Thread.Sleep(1000);
-            Assert.Equal(0, (int)poolService.GetTmpSizeAsync().Result);
-            Assert.Equal(5, (int) poolService.GetPoolSize().Result);
-            Assert.Equal(4, (int)poolService.GetWaitingSizeAsync().Result);
-            Assert.Equal(0, (int)poolService.GetExecutableSizeAsync().Result);
+            Assert.Equal(0, (int) await poolService.GetTmpSizeAsync());
+            Assert.Equal(5, (int) await poolService.GetPoolSize());
+            Assert.Equal(4, (int) await poolService.GetWaitingSizeAsync());
+            Assert.Equal(0, (int) await poolService.GetExecutableSizeAsync());
             
             var tx6 = new Transaction
             {
@@ -119,9 +119,9 @@ namespace AElf.Kernel.Tests.TxPool
             };
             res = await poolService.AddTxAsync(tx6);
             Assert.True(res);
-            Assert.Equal(1, (int)poolService.GetTmpSizeAsync().Result);
-            Assert.Equal(4, (int)poolService.GetWaitingSizeAsync().Result);
-            Assert.Equal(0, (int)poolService.GetExecutableSizeAsync().Result);
+            Assert.Equal(1, (int) await poolService.GetTmpSizeAsync());
+            Assert.Equal(4, (int) await poolService.GetWaitingSizeAsync());
+            Assert.Equal(0, (int) await poolService.GetExecutableSizeAsync());
             
             await poolService.Stop();
             
@@ -135,69 +135,5 @@ namespace AElf.Kernel.Tests.TxPool
             Assert.False(res);
             
         }
-
-        
-        
-        [Fact]
-        public async Task IntergrationTest()
-        {
-            var pool = new TxMemPool.TxPool(new ChainContext(_smartContractZero, Hash.Generate()), TxPoolConfig.Default,
-                _accountContextService);
-            
-            var poolService = new TxPoolService(pool);
-            poolService.Start();
-            ulong queued = 0;
-            ulong exec = 0;
-            var tasks = new List<Task>();
-            int k = 0;
-            var threadNum = 10;
-            for (var j = 0; j < 10; j++)
-            {
-                var task = Task.Run(async () =>
-                {
-                    var sortedSet = new SortedSet<ulong>();
-                    var addr = Hash.Generate();
-                    var i = 0;
-                    while (i++ < 500)
-                    {
-                        var id = (ulong) new Random().Next(100);
-                        sortedSet.Add(id);
-                        var tx = new Transaction
-                        {
-                            From = addr,
-                            To = Hash.Generate(),
-                            IncrementId = id
-                        };
-
-                        await poolService.AddTxAsync(tx);
-                    }
-
-                    ulong c = 0;
-                    foreach (var t in sortedSet)
-                    {
-                        if (t != c)
-                            break;
-                        c++;
-                    }
-                    
-                    lock (this)
-                    {
-                        queued += (ulong) sortedSet.Count;
-                        exec += c;
-                        k++;
-                    }
-                });
-                tasks.Add(task);
-            }
-
-            Task.WaitAll(tasks.ToArray());
-            pool.QueueTxs();
-            await poolService.PromoteAsync();
-            Assert.Equal(k, threadNum);
-            Assert.Equal(exec, poolService.GetExecutableSizeAsync().Result);
-            Assert.Equal(queued - exec, poolService.GetWaitingSizeAsync().Result);
-            
-        }
-        
     }
 }
